.requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
django-admin-thumbnails==0.2.8
django-richtextfield==1.6.1
django-tinymce==3.6.1
pillow==10.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
